Update Playlist.cs
Added Clear method which resets the state of the playlist.

Added the ability to move back or forward by an number of iterations rather than just once.

Directly editing the Position property no longer triggers the PositionChanged event, that event is now only trigger when a MoveNext or Move Previous is executed.