Create WorkPump.cs
The WorkPump class handles a work queue on a dedicated thread, the specification of that thread can be customized by the caller when `Run` is called, such as the ThreadPriority and the Name.

The work loop uses an efficient WaitHandle system so that the thread doesn't uselessly spin when waiting for work.

Defines the IWorkItem which callers can implement to pass their work to the pump, their work must initiate on a call to `Invoke()`.