Update AudioServer.cs
Fixed an oversight whereby the song would cut around 4 seconds early. This was due to the fact that once samples had finished being sent, the server would send the StopAudio command as it wrongly assumed that because it had stopped sending the samples that the clients would have finished using them, this is obviously not the case as there are 4 seconds of samples sent prior to playing. I have fixed this oversight by adding a 4 second delay after the samples have finished sending.

This whole class does need a re-write in regard to the Playlist/Queue system.