Drastically improved playlist and audio serving system.
The audio server was doing too much in terms of managing the Playlist, following the SRP it should only have one real core function. The Audio Server now just plays or stops audio and emits updates to listeners via the PlaybackStarted and PlaybackStoppedNaturally events.

The playlist management is now moved up to the Host Form, this enabled me to fix a bug (which was basically a work around) whereby when you clicked an item, the item before it would be selected briefly before your item was then selected.

The Playlist class has been improved in the sense that it now also has one responsibility, the Pause and Stop states are now removed entirely.