Update AudioServer.cs
Added error handling if the AudioFileReader cannot read the audio file.

The Play method now uses a AboveNormal thread priority thread rather than a thread pool thread as it is a long running task and so is not suitable for the ThreadPool, additionally it needs to compete at the same priority as the Client which is handling Ping/Play commands at AboveNormal priority too.