Create TintableButton
Custom control TintableButton fills in missing functionality from the original Button class. You can now specify a tint which will be applied to the BackgroundImage of the button.

I have also added a Checked state and a Checked Tint colour. Although this does not technically fall under the remit of the TintableButton it is needed in the GUI.