Create AudioRerouter.cs
This part of the program allows the user to broadcast a Wasapi device's output rather than a audio file.

The main use case this has been optimized for is as follows:
Create a virtual audio cable (using Virtual Audio Cable), have Spotify output to that virtual cable, have the Audio Rerouter listen to the virtual cable, which will broadcast the Spotify music to all clients!