Updated HostForm.cs
Added a new ChangeGUIState function which takes a new GUIState enum and transforms the UI into that state. Currently it is only used to enable/disable certain buttons as the state changes but could easily be expanded.

The audio re-router no longer shows as a Dialog Box, as this would cause the HostForm GUI to lock up. Instead the audio rerouter is an independent form with its IWin32WindowOwner set to the HostForm, this way the Host form can still be interacted with. The Host form disables playlist controls whilst the Audio Rerouter is open.