Re-wrote the code that sends clients samples periodically.
The while loop now keeps track of how many bytes it has sent and theoretically how many have been consumed (by multiplying the time elapsed in milliseconds by the bytes per millisecond) and subtracting that each loop.

When the it is estimated that the client(s) have less than 2 seconds of audio, the method will send out 4 seconds (or as many that are available on the reader).