Fixed threading race condition.
Fixed a race condition in the work loop whereby if the pump was instructed to `Stop()` exactly prior to the work queue empty check, the `_waitForWorkSignal` would be reset and the work loop would stay in limbo waiting for the `_waitForWorkSignal` to be set again, the way out of limbo would be to call `Stop` again.

There is now a check on the `_stopPump` variable right before the `_waitForWorkSignal` is reset, meaning it won't reset if `Stop()` has been called.