Update AudioServer.cs
Fixes #7
When resuming from a specific byte position, the track "playing time watch" would start from Zero, and wait until it was greater than the track length, this means that if you resume from 30 seconds, the "playing time watch" would start from zero and so the host would not move to the next song until 30 seconds after the song had actually finished. This has been fixed now, the "playing time watch" is now no longer a stop watch and instead uses a Unix Time method to calculate progressed time, the time progressed is offset by the resume byte position (Byte `Offset / Bytes Per Second) * Milliseconds`

Fixes #6
Previously the track position that was exposed to the Host was actually the buffered position (which is around 4 seconds ahead of the actual track position), the host would use this value to determine where it had been paused (4 seconds ahead), now, we use the track position which has been exposed via the fix above this one, and multiply the position by the Byte Rate, this means the host can get the exact bytes that it was paused on.