Update AudioServer.cs
Drastic improvements to the AudioServer.

Subscribers may now be notified when playback begins using the PlaybackStarted event trigger, this event contains `TrackInformation` which the host can use to make decisions on how to setup the UI, such as track bars etc.

Expanded upon making the API more intuative, the StartOffset and PreBuffer size are now no longer int/long parameters, they are TimeSpan, previously callers would have to know all about the current track and the insides of the AudioServer in order to reposition the stream etc, this has been abstracted away now, the Play method does all the calculation for the caller, the caller just provides a TimeSpan of where they would like to reposition to and a TimeSpan of how many seconds they would like to prebuffer.

Fixed some inefficiencies in the PlaySong method, previously if the user were scrubbing, Stop would be called and PlaySong would be recalled, this is still the same however PlaySong is more intelligent, instead of immediately disposing of the FileStream as soon as Stop is called, callers can set the `disposeOfFile` parameter to false, which will leave the FileStream in `_lastLoadedFile` until the next call to PlaySong, PlaySong will check whether there is already an open FileStream before trying to create a whole new AudioFileReader and having to reacquire a new file handle etc.